feat: sniper implementation

+++ SkinSniper/Services/Telegram/Entities/TelegramItem.cs
@@ -4,9 +4,9 @@ namespace SkinSniper.Services.Telegram.Entities
 {
     internal class TelegramItem
     {
-        public float BuffPrice { get; set; }
-        public float ListingPrice { get; set; }
-        public float Profit { get; set; }
+        public decimal BuffPrice { get; set; }
+        public decimal ListingPrice { get; set; }
+        public decimal Profit { get; set; }
         public bool Sniped { get; set; }
         public long TimeTaken { get; set; }
     }
